readded robot model, connected game to react router
